refactor(p2): extract ModelSettingsForm bundle into sibling page (stage 2 of 3)
ModelSettingsForm + 7 helpers (parseContextWindowInput,
Int.formatContextWindowInput, ModelTypeSelector, ModelModalitySelector,
ModalAbilitySelector, BuiltInToolsSettings, ProviderOverrideSettings)
moved from SettingProviderDetailPage.kt (4 non-contiguous ranges in
the Stage-1-trimmed coordinator: L249-277 + L460-657 + L1004-1148 +
L1340-1564) into new sibling SettingProviderModelSettingsForm.kt.

Byte-equal: bodies and signatures unchanged.
Visibility widened: ModelSettingsForm private -> internal (still called
by ModelCard at coord L1351 and AddModelButton at coord L933, both
remaining in coordinator for Stage 3).
ModalAbilitySelector kept its pre-existing public visibility byte-equal
(no tightening — confirmed 0 cross-file callers).
6 internal helpers (parseContextWindowInput, formatContextWindowInput,
ModelTypeSelector, ModelModalitySelector, BuiltInToolsSettings,
ProviderOverrideSettings) stay private in sibling.

Coordinator strip: 1564 -> 943 (-621 with import cleanup).
Imports trimmed: 19 this-refactor orphans (rememberScrollState,
KeyboardOptions, verticalScroll, MultiChoiceSegmentedButtonRow,
SecondaryTabRow, SegmentedButton, SegmentedButtonDefaults,
SingleChoiceSegmentedButtonRow, Switch, Tab, Color, KeyboardType,
BuiltInTools, Modality, ModelAbility, ModelType, CustomBodies,
CustomHeaders, ProviderConfigure). All confirmed: >=1 body ref in
67e7e442 baseline, 0 in trimmed coordinator.

Pre-existing dead imports preserved per CLAUDE.md §3: Connect, height,
LinearWavyProgressIndicator, CoroutineScope, TextGenerationParams,
defaultCodexOAuthModelList, UIMessage, ModelSelector, extendColors,
UiState.

getValue/setValue/plus kept — implicit usage in remaining coord:
2 val-by + 3 var-by sites + 1 PaddingValues+PaddingValues at L309.

Junction handling: 3 junctions in coordinator after multi-range strip
collapsed from double-blank to single (Python pass). EOF normalized to
single \n.

Baseline 3/467 unit tests preserved.
